rust: reorganize code and split library to smaller parts
